refactor: modularise dashboard components - move Dash app & server to `src/core.py` - move opioid graph related variables to `src/modules/components/opioid.py` - move deaths graph related variables to `src/modules/components/deaths.py` - move demo graph related variables to `src/modules/components/demo.py` - move sidebar related variables to `src/modules/components/sidebar.py` - rename `cards.py` to `agg_vals.py` - minor import statement formattings
