--- src/modules/components/__init__.py
@@ -1,3 +1,0 @@
-from .footer import footer
-from .footnote import footnote
-from .cards import *
