--- src/modules/components/cards.py
@@ -1,18 +1,0 @@
-from dash import html
-import dash_bootstrap_components as dbc
-
-
-def create_card(title, value, id_value):
-    return dbc.Card(
-        [
-            html.H4(title, className="card-title"),
-            html.P(value, className="card-value", id=id_value),
-        ],
-        body=True,
-    )
-
-
-death_card = create_card("Cumulative Deaths from All Drugs", 0, "death-value")
-death_rate_card = create_card("Average Death Rate", 0, "death-rate-value")
-percentage_card = create_card("Percentage Young Adult Deaths", 0, "percentage-value")
-fold_change_card = create_card("Fold Change", 0, "fold-change-value")
