+++ src/modules/components/footer.py
@@ -1,4 +1,5 @@
 from dash import html, dcc
+
 from ..utils import REPO_URL, get_repo_last_updated_time
